requirements.txt
fastapi==0.97.0
uvicorn[standard]==0.22.0

SQLAlchemy==2.0.16
sqlalchemy-utils==0.41.1

# for mypy
sqlalchemy-stubs==0.4

ruff
pydantic[email]

python-dotenv==1.0.0

# PostgreSQL database adapter
psycopg2-binary==2.9.6

#database migrations
alembic==1.11.1

# python linter
ruff==0.0.272

# type checker
mypy==1.3.0

# test packages
pytest==7.3.2
pytest-cov==4.1.0
factory-boy==3.2.1

# required for TestClient
httpx==0.24.1

